batch: Split hunk ownership translation
Live-hunk selection translation now resides in a hunk-focused batch module.
Shared LineEntry ownership primitives sit in an adapter module used by both
translation paths.

Command and source-refresh callers import the hunk translator directly.
Direct display-line translation remains under ownership_translation, which
leaves stale-source line handling in its own module.

This commit keeps hunk scan state, replacement-run coupling, and old-line
baseline reference construction out of the direct line-selection path.

Command behavior is unchanged, while internal ownership flow now names the
live-hunk boundary explicitly.
